Ordering in group lexicon. Ranking of interpretations in CS.
For a given key (head), the groups are ordered by the number of tokens
and features, and only the first is permitted to match. The evaluation
function for search states in constraint satisfaction now prefers
groups with lower indices, which matters for cached analyses that take
frequency into account. Noun genitive constructions (“casa de N”) in
noun group lexicon.