Fixed between-group dependencies with a new sentence-level constraint.
The problem came up when a group with a cat node needed to merge that
node with a lexical node in another group, but there was a competing
group for that node because the source token in that position was
ambiguous. For example, <hacia $n> in the context of “partes”, which
could be a noun (right) or a verb (wrong). The solution is to force the
group for the correct interpretation to succeed with a new constraint
the enforces the dependency between the higher-level group (the one
with the cat node) and the lower-level group (the one with the lexical
node). This involved the creation of a new constraint type,
DependencySelection, with parameters for a selection variable ($groups
in this case) and a list of determined set variables with indices of
elements (here groups) that each element (group) depends on. See
constraint.py and sentence.py.