Lots of changes to constraint satisfaction.
For selection of distribution variable: if there are group conflicts,
one of these values for $groups is selected. Otherwise if
$covered_snodes is not determined, a value for it is selected,
preferably one of an SNode that is the site of a merger. SearchState
(DState) evaluation uses a measure of the number of covered SNodes and
the average size of groups. Also in progress is changing the way
Spa->Grn relative clauses are handled (in groups rather than morphosys
patterns.