More stuff to deal with ambiguity.
First, ambiguous analyses of source words were leading to matching of
morphosyns for one analysis and elimination of the other, even if it
was the preferred analysis. Now this causes the sentence to be copied
(as an altsyn) so that the version without the morphosyn applied is
also returned. Second, more than one group with the same key may now be
considered, but not if the position and group length are the same.
(That is, the order groups are examined still matters.) Finally, there
were some changes in the categories for Spa pronouns and articles.