Recording translations now works.
With the "recordar" option, a Translation object is created (for a Text
object that is also created if it doesn't exist), and for each sentence in the Translation, a TraSeg object is created, storing the text of the sentence translation.
TextDB.align(Translation) returns a list of source, target string pairs.